Implement Resizable Scroll Bar for Animation Player
Changed the scroll bar in Animation Player to a resizable scrollbar
with handles at the edges. Pulling the handles change the zoom of
the timeline. When one handle is pulled the other one stays in
the same value.

Make other edge of scroll bar stay in place when dragging an edge

Fix the max timeline length to 120%(?) of the animation length

Rename some things to improve clarity

Co-authored-by: Miguel Jin <miguel.jin@tecnico.ulisboa.pt>
Co-authored-by: Malcolm Anderson <malcolminyo@gmail.com>